# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: payment
    description: Payment operations via Ashtech Pay
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /countries:
    get:
      operationId: getCountries
      tags: [payment]
      summary: Get available countries and operators
      description: Returns the list of active countries and their Mobile Money operators
      responses:
        "200":
          description: List of countries and operators
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountriesResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /fees:
    get:
      operationId: getFees
      tags: [payment]
      summary: Get current fee rates
      description: Returns the platform fee rates in real time. Use credited_amount from /collect for the exact net amount per transaction.
      responses:
        "200":
          description: Fee rates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeesResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Gateway error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collect:
    post:
      operationId: initiatePayment
      tags: [payment]
      summary: Initiate a Mobile Money payment
      description: |
        Initiates a payment and returns the transaction status.
        The platform fees are deducted automatically — credited_amount is the net amount credited.

        Flow detection:
        - 202 + flow="wave": Wave payment — display wave_url as button/QR code
        - 202 (no wave): USSD Push — wait for webhook, client validates on phone
        - 400 + error="otp_required" + ussd_code: OTP USSD (Orange BF) — client dials USSD code then sends OTP
        - 400 + error="otp_required" + ussd_code=null: OTP SMS — client receives SMS OTP, re-submit with otp field
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectRequest"
      responses:
        "202":
          description: Payment initiated (USSD Push or Wave)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectResponse"
        "400":
          description: OTP required or bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequiredResponse"
        "403":
          description: Forbidden — transaction does not belong to this account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found — transaction does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Unprocessable — country or operator not supported / wrong currency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited — too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Gateway error — operator network rejected the payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transaction/{id}:
    get:
      operationId: getTransaction
      tags: [payment]
      summary: Get transaction status
      description: Returns the current status of a transaction by its transaction_id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "403":
          description: Forbidden — transaction does not belong to this account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /webhook:
    post:
      operationId: handleWebhook
      tags: [payment]
      summary: Receive payment webhook events
      description: |
        Called by Ashtech Pay when a transaction reaches a final state.
        Always respond HTTP 200 immediately, then process business logic.
        The amount field is the net amount (after fees). total_amount is the gross collected amount.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookEvent"
      responses:
        "200":
          description: Webhook received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                required:
                  - received
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Country:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        currency:
          type: string
        operators:
          type: array
          items:
            type: string
      required:
        - name
        - code
        - currency
        - operators
    CountriesResponse:
      type: object
      properties:
        countries:
          type: array
          items:
            $ref: "#/components/schemas/Country"
      required:
        - countries
    FeesResponse:
      type: object
      description: Platform fee rates returned by Ashtech Pay
      properties:
        fees:
          type: array
          items:
            $ref: "#/components/schemas/FeeEntry"
      required:
        - fees
    FeeEntry:
      type: object
      properties:
        operator:
          type: string
        country_code:
          type: string
        fee_type:
          type: string
          description: "percentage or fixed"
        fee_value:
          type: number
          description: "Fee rate (e.g. 2.5 for 2.5%) or fixed amount"
        min_fee:
          type: number
        max_fee:
          type: number
      required:
        - operator
        - country_code
        - fee_type
        - fee_value
    CollectRequest:
      type: object
      properties:
        amount:
          type: number
          description: Gross amount to collect
        currency:
          type: string
          description: "Country currency (XAF, XOF, GNF, CDF…)"
        phone:
          type: string
          description: Payer phone number
        operator:
          type: string
          description: Exact operator name from /countries (e.g. MTN Mobile Money)
        country_code:
          type: string
          description: "ISO country code (CM, SN, CI…)"
        customer_name:
          type: string
          description: Payer full name (optional, not forwarded to Ashtech Pay)
        reference:
          type: string
          description: Your unique order reference
        otp:
          type: string
          description: OTP code if required (see 400 otp_required response)
        notify_url:
          type: string
          description: Webhook URL to receive the payment result
      required:
        - amount
        - currency
        - phone
        - operator
        - country_code
    CollectResponse:
      type: object
      properties:
        transaction_id:
          type: string
        reference:
          type: string
        status:
          type: string
          description: "pending | success | failed"
        amount:
          type: number
          description: Gross collected amount
        credited_amount:
          type: number
          description: Net amount credited after platform fees
        fee_amount:
          type: number
          description: Platform fee deducted
        currency:
          type: string
        flow:
          type: string
          description: "Payment flow: wave | ussd_push | otp_sms | otp_ussd"
        wave_url:
          type: string
          description: Wave payment URL (only present when flow=wave)
      required:
        - transaction_id
        - status
        - amount
        - currency
    OtpRequiredResponse:
      type: object
      properties:
        error:
          type: string
          description: "Always otp_required"
        message:
          type: string
        ussd_code:
          type: string
          nullable: true
          description: "USSD code to dial (OTP USSD flow, Orange BF only). Null for OTP SMS flow."
      required:
        - error
        - message
    TransactionResponse:
      type: object
      properties:
        transaction_id:
          type: string
        reference:
          type: string
        status:
          type: string
          description: "pending | success | failed"
        amount:
          type: number
          description: Gross collected amount
        credited_amount:
          type: number
          description: Net amount credited after fees
        fee_amount:
          type: number
          description: Platform fee deducted
        currency:
          type: string
        phone:
          type: string
        created_at:
          type: string
          format: date-time
        confirmed_at:
          type: string
          format: date-time
      required:
        - transaction_id
        - status
        - amount
        - currency
    WebhookEvent:
      type: object
      description: |
        Event sent by Ashtech Pay when a transaction reaches a final state.
        Events: payment.completed, payment.failed, payout.completed, payout.failed
        Note: amount = net after fees, total_amount = gross collected
      properties:
        event:
          type: string
          description: "payment.completed | payment.failed | payout.completed | payout.failed"
        transaction_id:
          type: string
        reference:
          type: string
        status:
          type: string
        amount:
          type: number
          description: Net amount after platform fees
        total_amount:
          type: number
          description: Gross collected amount (before fees)
        currency:
          type: string
        phone:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - event
        - transaction_id
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message