# lib/api-zod/src/generated/types/collectRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CollectRequest {
  /** Gross amount to collect */
  amount: number;
  /** Country currency (XAF, XOF, GNF, CDF…) */
  currency: string;
  /** Payer phone number */
  phone: string;
  /** Exact operator name from /countries (e.g. MTN Mobile Money) */
  operator: string;
  /** ISO country code (CM, SN, CI…) */
  country_code: string;
  /** Payer full name (optional, not forwarded to Ashtech Pay) */
  customer_name?: string;
  /** Your unique order reference */
  reference?: string;
  /** OTP code if required (see 400 otp_required response) */
  otp?: string;
  /** Webhook URL to receive the payment result */
  notify_url?: string;
}