# lib/api-zod/src/generated/types/collectResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CollectResponse {
  transaction_id: string;
  reference?: string;
  /** pending | success | failed */
  status: string;
  /** Gross collected amount */
  amount: number;
  /** Net amount credited after platform fees */
  credited_amount?: number;
  /** Platform fee deducted */
  fee_amount?: number;
  currency: string;
  /** Payment flow: wave | ussd_push | otp_sms | otp_ussd */
  flow?: string;
  /** Wave payment URL (only present when flow=wave) */
  wave_url?: string;
}