# lib/api-zod/src/generated/types/feeEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FeeEntry {
  operator: string;
  country_code: string;
  /** percentage or fixed */
  fee_type: string;
  /** Fee rate (e.g. 2.5 for 2.5%) or fixed amount */
  fee_value: number;
  min_fee?: number;
  max_fee?: number;
}