# lib/api-zod/src/generated/types/feesResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeeEntry } from './feeEntry';

/**
 * Platform fee rates returned by Ashtech Pay
 */
export interface FeesResponse {
  fees: FeeEntry[];
}