# lib/api-zod/src/generated/types/otpRequiredResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OtpRequiredResponse {
  /** Always otp_required */
  error: string;
  message: string;
  /** USSD code to dial (OTP USSD flow, Orange BF only). Null for OTP SMS flow. */
  ussd_code?: string | null;
}