# lib/api-zod/src/generated/types/transactionResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TransactionResponse {
  transaction_id: string;
  reference?: string;
  /** pending | success | failed */
  status: string;
  /** Gross collected amount */
  amount: number;
  /** Net amount credited after fees */
  credited_amount?: number;
  /** Platform fee deducted */
  fee_amount?: number;
  currency: string;
  phone?: string;
  created_at?: Date;
  confirmed_at?: Date;
}